Extract large test and projection-ACP modules into submodules for file-size hygiene
Split three oversized source files into focused submodules using the
file+directory module resolution pattern (foo.rs + foo/bar.rs):

- trace.rs: Extract projection-ACP read-scope, binding discovery, and
  filtering logic (~1300 lines) to trace/projection_acp.rs.  The parent
  module retains dispatch, adapter-projection, and non-ACP tests.

- validate.rs: Extract the live_tests integration module (~1075 lines)
  to validate/live_tests.rs.  The parent module retains validation logic
  and pub(super) helper functions.

- request.rs: Extract the test module (~1353 lines) to request/tests.rs.
  The parent module retains request construction logic and the
  lean_vocab_test #[path] declaration.

All tests, types, and visibility modifiers preserved.  No behavioral
changes.
